Fix float_to_int16_altivec prototype to match float_to_int16's in dsputil.h (parameter 'len' is a long not an int). Patch by David Conrad % lessen42 A gmail P com %
Originally committed as revision 16451 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/ppc/float_altivec.c
@@ -184,7 +184,7 @@ float_to_int16_one_altivec(const float *src)
     return vec_packs(t0,t1);
 }
 
-static void float_to_int16_altivec(int16_t *dst, const float *src, int len)
+static void float_to_int16_altivec(int16_t *dst, const float *src, long len)
 {
     int i;
     vector signed short d0, d1, d;
